# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Philosophy 101 course API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  email: string;
  name: string;
}

export interface Student {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  intro: string | null;
  createdAt: string;
  isAdmin: boolean;
  accommodated: boolean;
  integrityAckAt: string | null;
}

export interface CurrentStudent {
  student: Student | null;
}

export interface Progress {
  completedModuleIds: string[];
  /** @nullable */
  intro: string | null;
}

export interface SaveIntroBody {
  intro: string;
}

/**
 * Lifecycle of the GPTZero check for this submission.
 */
export type SubmissionAiStatus =
  (typeof SubmissionAiStatus)[keyof typeof SubmissionAiStatus];

export const SubmissionAiStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Submission {
  id: number;
  moduleId: string;
  content: string;
  createdAt: string;
  /** Probability (0–1) that the submission was AI-generated, per GPTZero. Null while the check is pending or unavailable. */
  aiScore: number | null;
  /** GPTZero predicted class — typically "human", "ai", or "mixed". */
  aiClass: string | null;
  aiCheckedAt: string | null;
  /** Lifecycle of the GPTZero check for this submission. */
  aiStatus: SubmissionAiStatus;
}

export interface SubmissionOrNull {
  submission: Submission | null;
}

export interface CreateSubmissionBody {
  moduleId: string;
  content: string;
}

export interface TutorMessage {
  id: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface TutorConversation {
  id: number;
  moduleId: string;
  messages: TutorMessage[];
}

export interface SendTutorMessageBody {
  content: string;
}

export interface CritiqueAnswer {
  answer: string;
}