# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Philosophy 101 course API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login or register a student by email
 */
export const LoginBody = zod.object({
  email: zod.string(),
  name: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  intro: zod.string().nullable(),
  createdAt: zod.coerce.date(),
  isAdmin: zod.boolean(),
  accommodated: zod.boolean(),
  integrityAckAt: zod.union([zod.coerce.date(), zod.null()]),
});

/**
 * @summary Get current logged-in student
 */
export const GetCurrentStudentResponse = zod.object({
  student: zod.union([
    zod.object({
      id: zod.number(),
      email: zod.string(),
      name: zod.string(),
      intro: zod.string().nullable(),
      createdAt: zod.coerce.date(),
      isAdmin: zod.boolean(),
      accommodated: zod.boolean(),
      integrityAckAt: zod.union([zod.coerce.date(), zod.null()]),
    }),
    zod.null(),
  ]),
});

/**
 * @summary List completed modules and current intro
 */
export const GetProgressResponse = zod.object({
  completedModuleIds: zod.array(zod.string()),
  intro: zod.string().nullable(),
});

/**
 * @summary Save student self-introduction
 */
export const SaveIntroBody = zod.object({
  intro: zod.string(),
});

export const SaveIntroResponse = zod.object({
  completedModuleIds: zod.array(zod.string()),
  intro: zod.string().nullable(),
});

/**
 * @summary List all of the current student's submissions
 */
export const ListSubmissionsResponseItem = zod.object({
  id: zod.number(),
  moduleId: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
  aiScore: zod
    .union([zod.number(), zod.null()])
    .describe(
      "Probability (0–1) that the submission was AI-generated, per GPTZero. Null while the check is pending or unavailable.",
    ),
  aiClass: zod
    .union([zod.string(), zod.null()])
    .describe(
      'GPTZero predicted class — typically \"human\", \"ai\", or \"mixed\".',
    ),
  aiCheckedAt: zod.union([zod.coerce.date(), zod.null()]),
  aiStatus: zod
    .enum(["pending", "completed", "failed"])
    .describe("Lifecycle of the GPTZero check for this submission."),
});
export const ListSubmissionsResponse = zod.array(ListSubmissionsResponseItem);

/**
 * @summary Submit an assessment for a module
 */
export const CreateSubmissionBody = zod.object({
  moduleId: zod.string(),
  content: zod.string(),
});

/**
 * @summary Get current student's most recent submission for a module
 */
export const GetSubmissionForModuleParams = zod.object({
  moduleId: zod.coerce.string(),
});

export const GetSubmissionForModuleResponse = zod.object({
  submission: zod.union([
    zod.object({
      id: zod.number(),
      moduleId: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
      aiScore: zod
        .union([zod.number(), zod.null()])
        .describe(
          "Probability (0–1) that the submission was AI-generated, per GPTZero. Null while the check is pending or unavailable.",
        ),
      aiClass: zod
        .union([zod.string(), zod.null()])
        .describe(
          'GPTZero predicted class — typically \"human\", \"ai\", or \"mixed\".',
        ),
      aiCheckedAt: zod.union([zod.coerce.date(), zod.null()]),
      aiStatus: zod
        .enum(["pending", "completed", "failed"])
        .describe("Lifecycle of the GPTZero check for this submission."),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Get or create the tutor conversation for a module
 */
export const GetTutorConversationParams = zod.object({
  moduleId: zod.coerce.string(),
});

export const GetTutorConversationResponse = zod.object({
  id: zod.number(),
  moduleId: zod.string(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Send a message to the tutor (SSE stream of assistant response)
 */
export const SendTutorMessageParams = zod.object({
  moduleId: zod.coerce.string(),
});

export const SendTutorMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Generate a deliberately weak AI answer for the critique exercise
 */
export const GenerateCritiqueAnswerParams = zod.object({
  moduleId: zod.coerce.string(),
});

export const GenerateCritiqueAnswerResponse = zod.object({
  answer: zod.string(),
});