# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Philosophy 101 course API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  intro: string | null;
  createdAt: Date;
  isAdmin: boolean;
  accommodated: boolean;
  integrityAckAt: Date | null;
}