# lib/api-zod/src/generated/types/submissionOrNull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Philosophy 101 course API
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from "./submission";

export interface SubmissionOrNull {
  submission: Submission | null;
}