# lib/api-zod/src/generated/types/tutorConversation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Philosophy 101 course API
 * OpenAPI spec version: 0.1.0
 */
import type { TutorMessage } from "./tutorMessage";

export interface TutorConversation {
  id: number;
  moduleId: string;
  messages: TutorMessage[];
}